+++ test/message_pack_object_test.dart
@@ -3,17 +3,16 @@ import "dart:typed_data";
 import "package:msg_pck/msg_pck.dart";
 import "package:test/test.dart";
 
-class DummyClass with Message {
+class DummyClass with MessagePackObject {
   late int f1;
   late String f2;
   late Map<dynamic, dynamic> f3;
   late List<dynamic> f4;
   late Uint8List f5;
-  // late double f6; https://github.com/dart-lang/sdk/issues/53284
+  late double f6;
 
   @override
-  List<dynamic> get messagePackFields => <dynamic>[f1, f2, f3, f4, f5];
-  // List<dynamic> get messagePackFields => <dynamic>[f1, f2, f3, f4, f5, f6];
+  List<dynamic> get messagePackFields => <dynamic>[f1, f2, f3, f4, f5, f6];
 
   DummyClass();
 
@@ -23,7 +22,7 @@ class DummyClass with Message {
     f3 = items[2] as Map<dynamic, dynamic>;
     f4 = items[3] as List<dynamic>;
     f5 = items[4] as Uint8List;
-    // f6 = items[5] as double;
+    f6 = items[5] as double;
   }
 }
 
@@ -40,8 +39,8 @@ void main() {
         ..f2 = "Yo"
         ..f3 = <dynamic, dynamic>{"Hi": -3}
         ..f4 = <dynamic>["Hi", 11111, "Yo", 4]
-        ..f5 = Uint8List.fromList(<int>[1, 2, 3]);
-      // ..f6 = 0.01;
+        ..f5 = Uint8List.fromList(<int>[1, 2, 3])
+        ..f6 = 1.23;
 
       writer.writeMessage(myObject);
 
@@ -76,6 +75,11 @@ void main() {
             1,
             2,
             3,
+            202,
+            63,
+            157,
+            112,
+            164,
           ]),
         ),
       );
@@ -87,8 +91,8 @@ void main() {
         ..writeString("Yo")
         ..writeMap(<dynamic, dynamic>{"Hi": -3})
         ..writeArray(<dynamic>["Hi", 11111, "Yo", 4])
-        ..writeBinary(Uint8List.fromList(<int>[1, 2, 3]));
-      // ..writeDouble(0.01);
+        ..writeBinary(Uint8List.fromList(<int>[1, 2, 3]))
+        ..writeFloat(1.23);
 
       Uint8List data = writer.takeBytes();
 
@@ -102,7 +106,7 @@ void main() {
       expect(myObject.f3, equals(<dynamic, dynamic>{"Hi": -3}));
       expect(myObject.f4, equals(<dynamic>["Hi", 11111, "Yo", 4]));
       expect(myObject.f5, equals(Uint8List.fromList(<int>[1, 2, 3])));
-      // expect(myObject.f6, equals(0.01));
+      expect((myObject.f6 - 1.23).abs(), lessThan(0.0001));
     });
   });
 }