+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Change mixin name from `Message` to `MessagePackObject`.
+- Enable tests for `double`.
+
 ## 0.1.0
 
 - Initial version.